+++ doc/configuration/configuration.rst
@@ -4,7 +4,7 @@
 Dream Configuration
 ###################
 
-`Schema <https://github.com/lsst-ts/ts_dream/blob/python/lsst/ts/dream/config_schema.py>`_
+`Schema <https://github.com/lsst-ts/ts_dream/blob/python/lsst/ts/dream/csc/config_schema.py>`_
 
 Configuration files are located in the `ts_config_ocs repo <https://github.com/lsst-ts/ts_config_ocs>`_.
