+++ doc/developer-guide/developer-guide.rst
@@ -14,11 +14,11 @@ Dependencies
 
 These are the dependencies for the latest version of ts_dream:
 
-* ts_xml - v9.0.0
+* ts_xml - v9.1
 * `SAL <https://ts-sal.lsst.io>`_ - v5.1.1
-* ts_salobj - v6.4.1
-* ts_idl - v3.1.3
-* ts_tcpip - v0.3.0
+* ts_salobj - v6.5
+* ts_idl - v3.2
+* ts_tcpip - v0.3
 
 .. _API:
 
@@ -27,7 +27,7 @@ Dream API
 
 The content in this section is autogenerated from docstrings.
 
-.. automodapi:: lsst.ts.dream
+.. automodapi:: lsst.ts.dream.csc
     :no-main-docstr:
     :no-inheritance-diagram:
 
@@ -52,7 +52,7 @@ An alternative
     cd develop/ts_dream
     pip install -e .[dev]
     pre-commit install # install black hook
-    pytest --cov lsst.ts.dream -ra
+    pytest --cov lsst.ts.dream.csc -ra
 
 
 .. _Usage:
