+++ python/lsst/ts/dream/csc/config_schema.py
@@ -23,7 +23,7 @@ import yaml
 CONFIG_SCHEMA = yaml.safe_load(
     """
     $schema: http://json-schema.org/draft-07/schema#
-    $id: https://github.com/lsst-ts/ts_dream/blob/master/python/lsst/ts/dream/config_schema.py
+    $id: https://github.com/lsst-ts/ts_dream/blob/master/python/lsst/ts/dream/csc/config_schema.py
     # title must end with one or more spaces followed by the schema version, which must begin with "v"
     title: DREAM v1
     description: Schema for DREAM configuration files
