+++ python/lsst/ts/dream/csc/mock/mock_dream.py
