+++ setup.py
@@ -43,7 +43,7 @@ setuptools.setup(
     name="ts_dream",
     description="Dutch Rubin Enhanced Atmospheric Monitor",
     use_scm_version={
-        "write_to": "python/lsst/ts/dream/version.py",
+        "write_to": "python/lsst/ts/dream/csc/version.py",
         "write_to_template": scm_version_template,
     },
     setup_requires=["setuptools_scm", "pytest-runner"],
