+++ tests/test_dream_csc.py
@@ -33,7 +33,7 @@ logging.basicConfig(
 
 class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
     async def asyncSetUp(self) -> None:
-        self.srv = dream.mock.MockDream(host="0.0.0.0", port=0)
+        self.srv = dream.csc.mock.MockDream(host="0.0.0.0", port=0)
         await self.srv.start_task
         self.mock_port = self.srv.port
         self.writer = None
@@ -46,7 +46,7 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
         await self.srv.exit()
 
     def basic_make_csc(self, initial_state, config_dir, simulation_mode, **kwargs):
-        return dream.DreamCsc(
+        return dream.csc.DreamCsc(
             initial_state=initial_state,
             config_dir=config_dir,
             simulation_mode=simulation_mode,
@@ -79,7 +79,7 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
         ):
             await self.assert_next_sample(
                 self.remote.evt_softwareVersions,
-                cscVersion=dream.__version__,
+                cscVersion=dream.csc.__version__,
                 subsystemVersions="",
             )
